third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/response-legacystream.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script type="text/javascript">
var test = async_test("Test response of XMLHttpRequest with responseType set to 'legacystream' for various readyState.");

test.step(function()
{
    var xhr = new XMLHttpRequest;

    xhr.responseType = "legacystream";
    assert_equals(xhr.responseType, "legacystream", "xhr.responseType");

    assert_equals(xhr.readyState, xhr.UNSENT, "xhr.readyState");
    assert_equals(xhr.response, null, "xhr.response during UNSENT");

    var seenStates = [];

    xhr.onreadystatechange = test.step_func(function() {
        // onreadystatechange can be invoked multiple times in LOADING state.
        if (seenStates.length == 0 || xhr.readyState != seenStates[seenStates.length - 1])
            seenStates.push(xhr.readyState);

        switch (xhr.readyState) {
        case xhr.UNSENT:
            assert_unreached('Unexpected readyState: UNSENT');
            return;

        case xhr.OPENED:
            assert_equals(xhr.response, null, "xhr.response during OPENED");
            return;

        case xhr.HEADERS_RECEIVED:
            assert_equals(xhr.response, null, "xhr.response during HEADERS_RECEIVED");
            return;

        case xhr.LOADING:
            assert_not_equals(xhr.response, null, "xhr.response during LOADING");
            return;

        case xhr.DONE:
            assert_equals(xhr.status, 200, "xhr.status");

            // Check that we saw all states.
            assert_array_equals(seenStates,
                                [xhr.OPENED, xhr.HEADERS_RECEIVED, xhr.LOADING, xhr.DONE]);

            test.done();
            return;

        default:
            assert_unreached('Unexpected readyState: ' + xhr.readyState)
            return;
        }
    });

    xhr.open('GET', '../resources/test.ogv', true);
    xhr.send();
});
</script>
</body>
</html>
